Add TMUX attaching and YOLO alias
YOLO alias is for giggles, but TMUX attaching is actually useful.

When run, it will check if there is tmux sessions set by the $TMUX
environment variable, and if there is it will perform a TMUX attach
and if not it will start a new TMUX session.

The YOLO alias just commits everything and sets the message as
YOLO.