Update setup script
Use links instead of copying the files over entirely so that any
additions can just be added to the repo and not have to be coppied by
hand back to the repo from the working copy after testing was done.

Using '[[ ]]' compound expressions in bash since the '[ ]' single
bracket version has some odd behaviors that need special escaping. And
were breaking when just using ||.