Add gnome-bell script to configure bell in gnome-terminal
The gnome shell by default has a really annoying bell sound, but it can
be turned off and changed to a visual bell that just flashes the screen.

Using frame flash will only flash on the window of the application that
emitted the bell, rather than flashing the entire screen.